backend/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
django-cors-headers==4.2.0
pdfkit==1.0.0
weasyprint==52.5